requirements.txt
requests==2.21.0
pytz==2018.9
tqdm==4.30.0
Send2Trash==1.5.0
selenium==3.141.0
cfscrape==1.9.5
Pillow==5.4.1
PyQt5==5.12
beautifulsoup4==4.7.1
